Feat: Move code into separate main jsx file
